requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
supabase
python-multipart 